# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateTaskInput:
    properties:
      completed:
        example: false
        type: boolean
      title:
        example: Belajar Go
        type: string
    type: object
  handlers.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  handlers.TaskResponse:
    properties:
      completed:
        example: false
        type: boolean
      id:
        example: 1
        type: integer
      title:
        example: Belajar Go
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  handlers.UpdateTaskInput:
    properties:
      completed:
        example: true
        type: boolean
      title:
        example: Belajar Swagger
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Todo API menggunakan Go Gin PostgreSQL JWT
  title: Todo API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login menggunakan email dan password untuk mendapatkan JWT token
      parameters:
      - description: Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Membuat akun baru
      parameters:
      - description: Register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Register user baru
      tags:
      - Auth
  /tasks:
    get:
      description: Mengambil semua task milik user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.TaskResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get semua task
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Membuat task baru untuk user yang sedang login
      parameters:
      - description: Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Membuat task baru
      tags:
      - Task
  /tasks/{id}:
    delete:
      description: Menghapus task milik user yang sedang login
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Hapus task
      tags:
      - Task
    get:
      description: Mengambil satu task berdasarkan ID milik user yang sedang login
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task berdasarkan ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Mengubah task milik user yang sedang login
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - Task
securityDefinitions:
  BearerAuth:
    description: Masukkan JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"